requirements.txt
agno
tavily-python
ollama

